Add `ethercrab-wire` and `ethercrab-wire-derive` crates (#141)
To replace `packed_struct`, `num-enum` and `nom`.

These are experimental crates that have custom derives for
reading/writing from/to EtherCAT wire representations. They will
hopefully become more useful to end users in the future for things like
autoconfiguration from ESI files and such.

This is a split of #139 as that PR is huge and unwieldy due to high
doses of  scope creep