Refactor `Command` and internal PDU loop a bunch (#146)
No user visible changes.

This is kind of a WIP PR as a split of #139 so the code is quite messy,
but it does still work fine!

The goal is to eventually remove `SlaveClient` and some of the
indirection around it. This PR is a step towards that, but not the
finished article.