Check group SubDevice states in chunks of 16
This makes the state check 16x more efficient by sending a single
EtherCAT frame with state requests for (up to) 16 SubDevices in the
group.

It would be great if the chunk size could be computed based on the PDU
size specified by the user, but Rust isn't there yet. A chunk size of 16
was chosen quite arbitrarily as a balance between allowing small PDU
sizes but more efficient state checks.