Implement SDO Info Get Object Description List
To add this endpoint, I added two methods
(`SubDeviceRef::{sdo_info_object_description_list,
sdo_info_object_quantities}`) and two types
(`crate::{ObjectDescriptionListType, ObjectDescriptionListTypeCounts}`)
to the public interface.  I also made a modest change to the
`MailboxError` type.

The application I have in mind for this feature is to list out the
object dictionaries of every device on the bus to improve surveying,
troubleshoot, and plain ol' hacking.  You can see the beginning of this
in `/examples/sdo-info.rs`.  Some devices on the bus won't have the
necessary mailboxes to support this kind of probing.  To handle that
gracefully, the log messages that were produced at the same site as the
`NoMailbox` errors needed to be lofted to where that error was actually
problematic.  But to preserve the existing logs, the errors needed to
specify whether it was a read or write mailbox that was missing, so I
created those errors.

In principal I could have implemented this endpoint with a single method
and type, as that's what the standard describes, but I elected to make
the list quantity request separate since that would enable a more
expressive API for the user with less documentation lookups for which
index is which list type.

I plan to expand the `sdo-info` example with more SDO Info protocols in
the coming weeks.