Add a label to timeouts for better errors (#320)
* Add a label to timeouts for better errors

Up to this commit, timeout errors have been inspecific about which
timeout is being tripped.  This can be frustrating, as timeout errors
are user-facing and often require action by the user, either to adjust
timeout durations or to do deeper diagnosis on the hardware.

This creates better error messages by adding a label to all timeouts so
that they will report which particular timeout they tripped.

The other most notable change is that
`crate::timer_factory::TimeoutFuture` has had it's `.duration` field
added regardless of whether it's built with miri.  I needed to add
`TimeoutKind` to this struct no matter what and I felt it would be silly
to separate it from `Duration` just for miri's sake.

I've tested this on Beckhoff's EK1100, EL3104, EL4134, EL6224, EL1259,
EL6751, HBM's TIM EC, and Bosch-Rexroth's VT-HMC.  The latter also
exercised the PDU timeouts in addition to the regular timer factory
ones.